Add files via upload
As an Intern, I have created and completed the task  in 5-7  days approximately on  Jupyter Lab Notebook as per the task of Prodigy InfoTech i.e., DATA SCIENCE 
TASK 1. Task : Create a bar chart or histogram to visualize the distribution of a categorical or continuous variable, such as the distribution of ages or genders in a population.